Update README.md
Community isn't just about sharing your snacks, it's about sharing your laughs too! It's the spice in PE-PELONs meditation, the pep in his prayers. For us, it's no different. A robust community is where the chuckles are plenty and support is as common as mismatched socks.
So, hop on over like a frog on a lily pad and join the jamboree. Together, we'll turn every 'ah' into a 'haha', and find the zany zen in being part of something bigger. Because with community, every day is a chance to belly laugh with buddies and live it up!